+++ appveyor.yml
@@ -30,7 +30,7 @@ build:
 
 test_script:
   - vstest.console /logger:Appveyor "toofz.NecroDancer.Leaderboards.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Leaderboards.Tests.dll"
-  - vstest.console /logger:Appveyor "toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests.dll"
-  - vstest.console /logger:Appveyor "toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.Tests.dll"
-  - vstest.console /logger:Appveyor "toofz.NecroDancer.Leaderboards.Services.ReplayUpdate.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Leaderboards.Services.ReplayUpdate.Tests.dll"
+  - vstest.console /logger:Appveyor "toofz.NecroDancer.Leaderboards.Services.LeaderboardsService.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Leaderboards.Services.LeaderboardsService.Tests.dll"
+  - vstest.console /logger:Appveyor "toofz.NecroDancer.Leaderboards.Services.PlayersService.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Leaderboards.Services.PlayersService.Tests.dll"
+  - vstest.console /logger:Appveyor "toofz.NecroDancer.Leaderboards.Services.ReplaysService.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Leaderboards.Services.ReplaysService.Tests.dll"
   - vstest.console /logger:Appveyor "toofz.NecroDancer.Web.Api.Tests\bin\%CONFIGURATION%\toofz.NecroDancer.Web.Api.Tests.dll"
