+++ toofz.NecroDancer.Leaderboards.Services.PlayersService/PlayerSettings.cs
@@ -1,6 +1,6 @@
 ﻿using toofz.NecroDancer.Leaderboards.Services.Common;
 
-namespace toofz.NecroDancer.Leaderboards.Services.PlayerUpdate
+namespace toofz.NecroDancer.Leaderboards.Services.PlayersService
 {
     sealed class PlayerSettings : Settings
     {
