+++ toofz.NecroDancer.Leaderboards.Services.PlayersService/Properties/AssemblyInfo.cs
@@ -2,14 +2,14 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("toofz Player Service")]
+[assembly: AssemblyTitle("toofz Players Service")]
 [assembly: AssemblyDescription("Provides Steam player data.")]
 [assembly: AssemblyCompany("toofz")]
-[assembly: AssemblyProduct("toofz Player Service")]
+[assembly: AssemblyProduct("toofz Players Service")]
 [assembly: AssemblyCopyright("Copyright © Leonard Thieu 2015")]
 [assembly: ComVisible(false)]
 [assembly: AssemblyVersion("1.0.*")]
 
-[assembly: InternalsVisibleTo("toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.Tests")]
+[assembly: InternalsVisibleTo("toofz.NecroDancer.Leaderboards.Services.PlayersService.Tests")]
 
-[assembly: log4net.Config.XmlConfigurator(Watch = true, ConfigFile = "player-log.config")]
+[assembly: log4net.Config.XmlConfigurator(Watch = true, ConfigFile = "log.config")]
