+++ toofz.NecroDancer.Leaderboards.Services.PlayersService/toofz.NecroDancer.Leaderboards.Services.PlayersService.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{714C9B00-9392-4F2A-872C-3FE575A23D12}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>toofz.NecroDancer.Leaderboards.Services.PlayerUpdate</RootNamespace>
-    <AssemblyName>PlayerUpdate</AssemblyName>
+    <RootNamespace>toofz.NecroDancer.Leaderboards.Services.PlayersService</RootNamespace>
+    <AssemblyName>PlayersService</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -33,7 +33,7 @@
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <NoWarn>1591</NoWarn>
-    <DocumentationFile>bin\Release\PlayerUpdate.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\PlayersService.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\toofz.NecroDancer.Leaderboards.Services.Common\toofz.NecroDancer.Leaderboards.Services.Common.csproj">
@@ -104,10 +104,10 @@
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
-    <None Include="player-log.config">
+    <None Include="log.config">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-    <None Include="player-settings.json">
+    <None Include="settings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
