+++ toofz.NecroDancer.Leaderboards.Services.ReplaysService/Program.cs
@@ -3,7 +3,7 @@ using log4net;
 using Microsoft.ApplicationInsights.Extensibility;
 using toofz.NecroDancer.Leaderboards.Services.Common;
 
-namespace toofz.NecroDancer.Leaderboards.Services.PlayerUpdate
+namespace toofz.NecroDancer.Leaderboards.Services.ReplaysService
 {
     static class Program
     {
@@ -12,14 +12,14 @@ namespace toofz.NecroDancer.Leaderboards.Services.PlayerUpdate
         static void Main(string[] args)
         {
             Log.Debug("Initialized logging.");
-            var instrumentationKey = Environment.GetEnvironmentVariable("PlayersInstrumentationKey", EnvironmentVariableTarget.Machine);
+            var instrumentationKey = Environment.GetEnvironmentVariable("ReplaysInstrumentationKey", EnvironmentVariableTarget.Machine);
             if (instrumentationKey != null) { TelemetryConfiguration.Active.InstrumentationKey = instrumentationKey; }
             else
             {
-                Log.Warn($"The environment variable 'PlayersInstrumentationKey' is not set. Telemetry is disabled.");
+                Log.Warn($"The environment variable 'ReplaysInstrumentationKey' is not set. Telemetry is disabled.");
                 TelemetryConfiguration.Active.DisableTelemetry = true;
             }
-            Application.Run<WorkerRole, PlayerSettings>();
+            Application.Run<WorkerRole, ReplaySettings>();
         }
     }
 }
